// src/api/inventory/models/Priority.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Priority = {
    Low: 'LOW',
    Medium: 'MEDIUM',
    High: 'HIGH',
    Critical: 'CRITICAL'
} as const;
export type Priority = typeof Priority[keyof typeof Priority];


export function PriorityFromJSON(json: any): Priority {
    return PriorityFromJSONTyped(json, false);
}

export function PriorityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Priority {
    return json as Priority;
}

export function PriorityToJSON(value?: Priority | null): any {
    return value as any;
}